+++ src/olcs/rastersynchronizer.js
@@ -94,23 +94,25 @@ olcs.RasterSynchronizer.prototype.convertLayerToCesiumImageries = function(olLay
 /**
  * @inheritDoc
  */
-olcs.RasterSynchronizer.prototype.createSingleLayerCounterparts = function(olLayer) {
+olcs.RasterSynchronizer.prototype.createSingleLayerCounterparts = function(olLayerWithParents) {
+  const olLayer = olLayerWithParents.layer;
   const uid = ol.getUid(olLayer).toString();
   const viewProj = this.view.getProjection();
   const cesiumObjects = this.convertLayerToCesiumImageries(olLayer, viewProj);
   if (cesiumObjects) {
     const listenKeyArray = [];
-    listenKeyArray.push(olLayer.on(['change:opacity', 'change:visible'],
-        (e) => {
-          // the compiler does not seem to be able to infer this
-          goog.asserts.assert(cesiumObjects);
-          for (let i = 0; i < cesiumObjects.length; ++i) {
-            olcs.core.updateCesiumLayerProperties(olLayer, cesiumObjects[i]);
-          }
-        }));
+    [olLayerWithParents.layer].concat(olLayerWithParents.parents).forEach((olLayerItem) => {
+      listenKeyArray.push(olLayerItem.on(['change:opacity', 'change:visible'], () => {
+        // the compiler does not seem to be able to infer this
+        goog.asserts.assert(cesiumObjects);
+        for (let i = 0; i < cesiumObjects.length; ++i) {
+          olcs.core.updateCesiumLayerProperties(olLayerWithParents, cesiumObjects[i]);
+        }
+      }));
+    });
 
     for (let i = 0; i < cesiumObjects.length; ++i) {
-      olcs.core.updateCesiumLayerProperties(olLayer, cesiumObjects[i]);
+      olcs.core.updateCesiumLayerProperties(olLayerWithParents, cesiumObjects[i]);
     }
 
     // there is no way to modify Cesium layer extent,