+++ CHANGES.md
@@ -3,6 +3,12 @@
 
 # v 1.32
 
+* Breaking changes
+  * Changed AbstractSynchronizer.createSingleLayerCounterparts Plugin API
+    in existing plugins the function has to be modified.
+    It now takes an object {layer: !ol.layer.Base, parents: Array<ol.layer.Group>}.
+    The parents are ordered with the first parent coming first and the last parent coming last.
+
 * Changes
   * Allow features to specify a dedicated geometry for rendering in Cesium
     using the `olcs.3d_geometry` property.
@@ -13,6 +19,7 @@
     initialization of an application. olcs.contrib.Manager is an abstract class
     which must be extended before use.
   * Added calcResolutionForDistance and calcDistanceForResolution functions to Camera API
+  * Added ol.layer.Group synchronization
 
 # v 1.31 - 2017-09-06